Final project 8th module

+++ src/Main.java
@@ -6,8 +6,8 @@ public class Main {
         File f2 = new File("C:\\Final8", "Text2.txt");
        boolean doesExist1 = f1.exists();
        boolean doesExist2 = f2.exists();
-        double f1Size = f1.length();
-        double f2Size = f2.length();
+        long f1Size = f1.length();
+        long f2Size = f2.length();
        if(doesExist1 == true) System.out.println("File's " +f1.getName()+ " size is " +f1Size+ " bytes");
        else System.out.println("File " +f1.getName()+ " doesn't exist");
         if(doesExist2 == true) System.out.println("File's " +f2.getName()+ " size is " +f2Size+ " bytes");
